history save_and_get api response updated

+++ Fastapi-backend/Wellnessapp/app/schemas/history.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from pydantic import BaseModel
 from datetime import datetime
 
@@ -12,7 +13,7 @@ class MealResponse(BaseModel):
     history_id: int
     meal_type_name: str
     category_name: str
-    food_kcal: float
+    food_kcal: Decimal
     food_car: int
     food_prot: int
     food_fat: int
