+++ Fastapi-backend/Wellnessapp/app/schemas/recommend.py
@@ -1,13 +1,14 @@
+from decimal import Decimal
 from pydantic import BaseModel
 from typing import Optional
 from datetime import datetime
 
 class RecommendBase(BaseModel):
     user_id: int
-    rec_kcal: float
-    rec_car: float
-    rec_prot: float
-    rec_fat: float
+    rec_kcal: Decimal
+    rec_car: Decimal
+    rec_prot: Decimal
+    rec_fat: Decimal
 
 class RecommendCreate(RecommendBase):
     pass
